BREAKING(path/unstable): move unstable overload of `dirname` to `unstable-dirname`
