+++ path/deno.json
@@ -38,11 +38,13 @@
     "./posix/resolve": "./posix/resolve.ts",
     "./posix/to-file-url": "./posix/to_file_url.ts",
     "./posix/to-namespaced-path": "./posix/to_namespaced_path.ts",
+    "./posix/unstable-dirname": "./posix/unstable_dirname.ts",
     "./relative": "./relative.ts",
     "./resolve": "./resolve.ts",
     "./to-file-url": "./to_file_url.ts",
     "./to-namespaced-path": "./to_namespaced_path.ts",
     "./types": "./types.ts",
+    "./unstable-dirname": "./unstable_dirname.ts",
     "./windows": "./windows/mod.ts",
     "./windows/basename": "./windows/basename.ts",
     "./windows/common": "./windows/common.ts",
@@ -62,6 +64,7 @@
     "./windows/relative": "./windows/relative.ts",
     "./windows/resolve": "./windows/resolve.ts",
     "./windows/to-file-url": "./windows/to_file_url.ts",
-    "./windows/to-namespaced-path": "./windows/to_namespaced_path.ts"
+    "./windows/to-namespaced-path": "./windows/to_namespaced_path.ts",
+    "./windows/unstable-dirname": "./windows/unstable_dirname.ts"
   }
 }
