+++ path/unstable_dirname.ts
@@ -2,27 +2,31 @@
 // This module is browser compatible.
 
 import { isWindows } from "./_os.ts";
-import { dirname as posixDirname } from "./posix/dirname.ts";
-import { dirname as windowsDirname } from "./windows/dirname.ts";
+import { dirname as posixDirname } from "./posix/unstable_dirname.ts";
+import { dirname as windowsDirname } from "./windows/unstable_dirname.ts";
 
 /**
- * Return the directory path of a path.
+ * Return the directory path of a file URL.
+ *
+ * @experimental **UNSTABLE**: New API, yet to be vetted.
  *
  * @example Usage
  * ```ts
- * import { dirname } from "@std/path/dirname";
+ * import { dirname } from "@std/path/unstable-dirname";
  * import { assertEquals } from "@std/assert";
  *
  * if (Deno.build.os === "windows") {
  *   assertEquals(dirname("C:\\home\\user\\Documents\\image.png"), "C:\\home\\user\\Documents");
+ *   assertEquals(dirname(new URL("file:///C:/home/user/Documents/image.png")), "C:\\home\\user\\Documents");
  * } else {
  *   assertEquals(dirname("/home/user/Documents/image.png"), "/home/user/Documents");
+ *   assertEquals(dirname(new URL("file:///home/user/Documents/image.png")), "/home/user/Documents");
  * }
  * ```
  *
  * @param path Path to extract the directory from.
  * @returns The directory path.
  */
-export function dirname(path: string): string {
+export function dirname(path: string | URL): string {
   return isWindows ? windowsDirname(path) : posixDirname(path);
 }
